+++ src/giskardpy/utils.py
@@ -29,6 +29,7 @@ import pylab as plt
 import pkg_resources
 
 from giskardpy.plugin import PluginBehavior, PluginBase
+from giskardpy.tfwrapper import kdl_to_pose, np_to_kdl
 
 
 @contextmanager
@@ -41,6 +42,7 @@ def suppress_stderr():
         finally:
             sys.stderr = old_stdout
 
+
 @contextmanager
 def suppress_stdout():
     devnull = os.open('/dev/null', os.O_WRONLY)
@@ -52,18 +54,23 @@ def suppress_stdout():
         os.dup2(old_stdout, 1)
         os.close(devnull)
 
+
 class NullContextManager(object):
     def __init__(self, dummy_resource=None):
         self.dummy_resource = dummy_resource
+
     def __enter__(self):
         return self.dummy_resource
+
     def __exit__(self, *args):
         pass
+
 
 class keydefaultdict(defaultdict):
     """
     A default dict where the key is passed as parameter to the factory function.
     """
+
     def __missing__(self, key):
         if self.default_factory is None:
             raise KeyError(key)
@@ -147,7 +154,8 @@ def closest_point_constraint_violated(closest_point_infos, tolerance=0.9):
     """
     for link_name, cpi_info in closest_point_infos.items():  # type: (str, ClosestPointInfo)
         if cpi_info.contact_distance < cpi_info.min_dist * tolerance:
-            print(u'collision constraints violated: {}'.format(cpi_info.link_a, cpi_info.link_b, cpi_info.contact_distance))
+            print(u'collision constraints violated: {}'.format(cpi_info.link_a, cpi_info.link_b,
+                                                               cpi_info.contact_distance))
             return True
     return False
 
@@ -271,6 +279,7 @@ def msg_to_list(thing):
                 thing.orientation.z,
                 thing.orientation.w]
 
+
 def create_path(path):
     if not os.path.exists(os.path.dirname(path)):
         try:
@@ -278,6 +287,7 @@ def create_path(path):
         except OSError as exc:  # Guard against race condition
             if exc.errno != errno.EEXIST:
                 raise
+
 
 def plot_trajectory(tj, controlled_joints, path_to_data_folder):
     """
@@ -311,10 +321,10 @@ def plot_trajectory(tj, controlled_joints, path_to_data_folder):
         ax1.plot(times, position, fmts[i], label=names[i])
         ax2.plot(times, velocities[i], fmts[i])
     box = ax1.get_position()
-    diff = abs(positions.max()-positions.min())*0.1
-    ax1.set_ylim(positions.min()-diff, positions.max()+diff)
-    diff = abs(velocities.max()-velocities.min())*0.1
-    ax2.set_ylim(velocities.min()-diff, velocities.max()+diff)
+    diff = abs(positions.max() - positions.min()) * 0.1
+    ax1.set_ylim(positions.min() - diff, positions.max() + diff)
+    diff = abs(velocities.max() - velocities.min()) * 0.1
+    ax2.set_ylim(velocities.min() - diff, velocities.max() + diff)
     ax1.set_position([box.x0, box.y0, box.width * 0.6, box.height])
     box = ax2.get_position()
     ax2.set_position([box.x0, box.y0, box.width * 0.6, box.height])
@@ -341,7 +351,10 @@ def resolve_ros_iris_in_urdf(input_urdf):
         output_urdf += u'\n'
     return output_urdf
 
+
 rospack = rospkg.RosPack()
+
+
 def resolve_ros_iris(path):
     if u'package://' in path:
         split = path.split(u'package://')
@@ -354,6 +367,7 @@ def resolve_ros_iris(path):
         return result
     else:
         return path
+
 
 def convert_dae_to_obj(path):
     path = path.replace(u'\'', u'')
@@ -370,6 +384,7 @@ def convert_dae_to_obj(path):
         return new_path
     return path
 
+
 def write_to_tmp(filename, urdf_string):
     """
     Writes a URDF string into a temporary file on disc. Used to deliver URDFs to PyBullet that only loads file.
@@ -385,6 +400,7 @@ def write_to_tmp(filename, urdf_string):
     with open(new_path, u'w') as o:
         o.write(urdf_string)
     return new_path
+
 
 def render_dot_tree(root, visibility_level=common.VisibilityLevel.DETAIL, name=None):
     """
@@ -425,6 +441,7 @@ def render_dot_tree(root, visibility_level=common.VisibilityLevel.DETAIL, name=N
     graph.write_png(filename_wo_extension + '.png')
     graph.write_svg(filename_wo_extension + '.svg')
 
+
 def generate_pydot_graph(root, visibility_level):
     """
     Generate the pydot graph - this is usually the first step in
@@ -437,6 +454,7 @@ def generate_pydot_graph(root, visibility_level):
     Returns:
         pydot.Dot: graph
     """
+
     def get_node_attributes(node, visibility_level):
         blackbox_font_colours = {common.BlackBoxLevel.DETAIL: "dodgerblue",
                                  common.BlackBoxLevel.COMPONENT: "lawngreen",
@@ -468,7 +486,8 @@ def generate_pydot_graph(root, visibility_level):
     graph.set_node_defaults(fontname='times-roman')
     graph.set_edge_defaults(fontname='times-roman')
     (node_shape, node_colour, node_font_colour) = get_node_attributes(root, visibility_level)
-    node_root = pydot.Node(root.name, shape=node_shape, style="filled", fillcolor=node_colour, fontsize=fontsize, fontcolor=node_font_colour)
+    node_root = pydot.Node(root.name, shape=node_shape, style="filled", fillcolor=node_colour, fontsize=fontsize,
+                           fontcolor=node_font_colour)
     graph.add_node(node_root)
     names = [root.name]
 
@@ -489,7 +508,8 @@ def generate_pydot_graph(root, visibility_level):
                 while proposed_dot_name in names:
                     proposed_dot_name = proposed_dot_name + "*"
                 names.append(proposed_dot_name)
-                node = pydot.Node(proposed_dot_name, shape=node_shape, style="filled", fillcolor=node_colour, fontsize=fontsize, fontcolor=node_font_colour)
+                node = pydot.Node(proposed_dot_name, shape=node_shape, style="filled", fillcolor=node_colour,
+                                  fontsize=fontsize, fontcolor=node_font_colour)
                 graph.add_node(node)
                 edge = pydot.Edge(root_dot_name, proposed_dot_name)
                 graph.add_edge(edge)
@@ -500,6 +520,7 @@ def generate_pydot_graph(root, visibility_level):
     add_edges(root, root.name, visibility_level)
     return graph
 
+
 def remove_outer_tag(xml):
     """
     :param xml:
@@ -507,7 +528,7 @@ def remove_outer_tag(xml):
     :return:
     :rtype: str
     """
-    return xml.split('>', 1)[1].rsplit('<',1)[0]
+    return xml.split('>', 1)[1].rsplit('<', 1)[0]
 
 
 def make_world_body_box(name=u'box', x_length=1, y_length=1, z_length=1):
@@ -520,6 +541,7 @@ def make_world_body_box(name=u'box', x_length=1, y_length=1, z_length=1):
     box.shape.dimensions.append(z_length)
     return box
 
+
 def make_world_body_sphere(name=u'sphere', radius=1):
     sphere = WorldBody()
     sphere.type = WorldBody.PRIMITIVE_BODY
@@ -528,15 +550,17 @@ def make_world_body_sphere(name=u'sphere', radius=1):
     sphere.shape.dimensions.append(radius)
     return sphere
 
+
 def make_world_body_cylinder(name=u'cylinder', height=1, radius=1):
     cylinder = WorldBody()
     cylinder.type = WorldBody.PRIMITIVE_BODY
     cylinder.name = str(name)
     cylinder.shape.type = SolidPrimitive.CYLINDER
-    cylinder.shape.dimensions = [0,0]
+    cylinder.shape.dimensions = [0, 0]
     cylinder.shape.dimensions[SolidPrimitive.CYLINDER_HEIGHT] = height
     cylinder.shape.dimensions[SolidPrimitive.CYLINDER_RADIUS] = radius
     return cylinder
+
 
 def make_urdf_world_body(name, urdf):
     wb = WorldBody()
@@ -544,6 +568,7 @@ def make_urdf_world_body(name, urdf):
     wb.type = wb.URDF_BODY
     wb.urdf = urdf
     return wb
+
 
 def is_iterable(qwe):
     try:
@@ -553,14 +578,7 @@ def is_iterable(qwe):
     return True
 
 def homo_matrix_to_pose(m):
-    p = Pose()
-    p.position.x = m[0,3]
-    p.position.y = m[1,3]
-    p.position.z = m[2,3]
-    p.orientation = Quaternion(*quaternion_from_matrix(m))
-    return p
-
-
+    return kdl_to_pose(np_to_kdl(m))
 
 def compare_version(version1, operator, version2):
     """
@@ -641,7 +659,7 @@ def rospkg_exists(name):
     name = name.replace(' ', '')
     version_list = name.split(',')
     version_entry1 = re.split('(==|>=|<=|<|>)', version_list[0])
-    package_name=version_entry1[0]
+    package_name = version_entry1[0]
     try:
         m = r.get_manifest(package_name)
     except Exception as e:
@@ -650,12 +668,14 @@ def rospkg_exists(name):
     if len(version_entry1) == 1:
         return True
     if not compare_version(version_entry1[2], version_entry1[1], m.version):
-        logwarn('found ROS package {installed_name}=={installed_version} but {r} is required}'.format(installed_name=package_name, installed_version=str(m.version), r=name))
+        logwarn('found ROS package {installed_name}=={installed_version} but {r} is required}'.format(
+            installed_name=package_name, installed_version=str(m.version), r=name))
         return False
     for entry in version_list[1:]:
         operator_and_version = re.split('(==|>=|<=|<|>)', entry)
         if not compare_version(operator_and_version[2], operator_and_version[1], m.version):
-            logwarn('found ROS package {installed_name}=={installed_version} but {r} is required}'.format(installed_name=package_name, installed_version=str(m.version), r=name))
+            logwarn('found ROS package {installed_name}=={installed_version} but {r} is required}'.format(
+                installed_name=package_name, installed_version=str(m.version), r=name))
             return False
 
     return True
@@ -680,4 +700,5 @@ def check_dependencies():
         except pkg_resources.DistributionNotFound as e:
             rospkg_exists(d)
         except pkg_resources.VersionConflict as e:
-            logwarn('found {version_f} but version {version_r} is required'.format(version_r=str(e.req), version_f=str(e.dist)))
+            logwarn('found {version_f} but version {version_r} is required'.format(version_r=str(e.req),
+                                                                                   version_f=str(e.dist)))
