+++ test/test_symengine_wrapper.py
@@ -18,8 +18,10 @@ from transforms3d.euler import euler2axangle
 from transforms3d.quaternions import quat2mat
 
 import giskardpy.symengine_wrappers as spw
+from giskardpy.tfwrapper import np_to_kdl, kdl_to_pose
+from giskardpy.utils import homo_matrix_to_pose, homo_matrix_to_pose2
 from utils_for_tests import limited_float, SMALL_NUMBER, unit_vector, quaternion, vector, \
-    pykdl_frame_to_numpy, lists_of_same_length, angle, compare_axis_angle, angle_positive
+    pykdl_frame_to_numpy, lists_of_same_length, angle, compare_axis_angle, angle_positive, compare_poses
 
 PKG = 'giskardpy'
